requirements.txt
numpy
opencv-python
requests
pyserial
pillow